make CLI more extensible
