+++ obsidian-cli/src/main.rs
@@ -1,246 +1,15 @@
 mod args;
+mod note;
 mod output;
-
-use std::env::current_dir;
-use std::io::{BufRead, IsTerminal};
+mod search;
+mod tags;
+mod utils;
 
 use clap::Parser;
 use color_eyre::eyre;
-use colored::Colorize;
-use obsidian_core::{LocatedTag, Note, Vault};
+use obsidian_core::Vault;
 
-use args::{
-    BacklinksArgs, Cli, Command, OutputFormat, RenameArgs, SearchArgs, SortOrder, TagsListArgs, TagsSearchArgs,
-    UpdateArgs,
-};
-
-fn modified_time(path: &std::path::Path) -> std::time::SystemTime {
-    std::fs::metadata(path)
-        .and_then(|m| m.modified())
-        .unwrap_or(std::time::SystemTime::UNIX_EPOCH)
-}
-
-fn sort_notes_by<T>(items: &mut [T], key: impl Fn(&T) -> &std::path::Path, sort: &SortOrder) {
-    match sort {
-        SortOrder::PathAsc => items.sort_by(|a, b| key(a).cmp(key(b))),
-        SortOrder::PathDesc => items.sort_by(|a, b| key(b).cmp(key(a))),
-        SortOrder::ModifiedAsc => items.sort_by_key(|a| modified_time(key(a))),
-        SortOrder::ModifiedDesc => items.sort_by_key(|b| std::cmp::Reverse(modified_time(key(b)))),
-    }
-}
-
-fn resolve_note_path(
-    vault: &Vault,
-    note_arg: &std::path::PathBuf,
-) -> eyre::Result<(std::path::PathBuf, std::path::PathBuf)> {
-    let (note_path, root) = if note_arg.is_absolute() {
-        (note_arg.clone(), vault.path.clone())
-    } else {
-        let cwd = current_dir()?;
-        let candidate1 = cwd.join(note_arg);
-        let candidate2 = vault.path.join(note_arg);
-        if candidate1.exists() {
-            (candidate1, cwd)
-        } else if candidate2.exists() {
-            (candidate2, vault.path.clone())
-        } else {
-            (note_arg.clone(), vault.path.clone())
-        }
-    };
-
-    if !note_path.exists() {
-        return Err(eyre::eyre!("note not found: {}", note_path.display()));
-    }
-
-    Ok((note_path, root))
-}
-
-fn cmd_search(vault: Vault, args: SearchArgs) -> eyre::Result<()> {
-    let mut query = vault.search();
-    for tag in &args.tag {
-        query = query.has_tag(tag);
-    }
-    for glob in &args.glob {
-        query = query.glob(glob);
-    }
-    for s in &args.content {
-        query = query.content_contains(s);
-    }
-    for t in &args.title_contains {
-        query = query.title_contains(t);
-    }
-    for alias in &args.alias {
-        query = query.has_alias(alias);
-    }
-    for s in &args.alias_contains {
-        query = query.alias_contains(s);
-    }
-    if let Some(r) = &args.regex {
-        query = query.content_matches(r);
-    }
-
-    let results = query.execute()?;
-    let mut notes: Vec<Note> = results.into_iter().filter_map(|r| r.ok()).collect();
-    sort_notes_by(&mut notes, |n| &n.path, &args.sort);
-
-    match args.format {
-        OutputFormat::Plain => output::print_search_plain(&notes, &vault.path),
-        OutputFormat::Json => output::print_search_json(&notes, &vault.path),
-    }
-    Ok(())
-}
-
-fn cmd_backlinks(vault: Vault, args: BacklinksArgs) -> eyre::Result<()> {
-    let (note_path, _) = resolve_note_path(&vault, &args.note)?;
-    let note = Note::from_path(&note_path)?;
-    let mut results = vault.backlinks(&note);
-    sort_notes_by(&mut results, |(n, _)| &n.path, &args.sort);
-
-    match args.format {
-        OutputFormat::Plain => output::print_backlinks_plain(&results, &vault.path),
-        OutputFormat::Json => output::print_backlinks_json(&results, &vault.path),
-    }
-    Ok(())
-}
-
-fn cmd_rename(vault: Vault, args: RenameArgs) -> eyre::Result<()> {
-    let (note_path, root) = resolve_note_path(&vault, &args.note)?;
-    let note = Note::from_path(&note_path)?;
-
-    let mut new_path = if args.new_path.is_absolute() {
-        args.new_path.clone()
-    } else {
-        root.join(&args.new_path)
-    };
-    if new_path.extension().and_then(|e| e.to_str()) != Some("md") {
-        new_path.set_extension("md");
-    }
-
-    if args.dry_run {
-        let preview = vault.rename_preview(&note, &new_path)?;
-        match args.format {
-            OutputFormat::Plain => output::print_rename_preview_plain(&preview, &vault.path),
-            OutputFormat::Json => output::print_rename_preview_json(&preview, &vault.path),
-        }
-    } else {
-        let renamed = vault.rename(&note, &new_path)?;
-        let rel = renamed.path.strip_prefix(&vault.path).unwrap_or(&renamed.path);
-        println!("{}", rel.display().to_string().cyan());
-    }
-    Ok(())
-}
-
-fn cmd_tags_search(vault: Vault, args: TagsSearchArgs) -> eyre::Result<()> {
-    let notes: Vec<Note> = vault.search().execute()?.into_iter().filter_map(|r| r.ok()).collect();
-
-    // A note tag matches a search term if it equals the term exactly or is a sub-tag of it
-    // (e.g. "workout/upper-body" matches search term "workout").
-    let tag_matches_search = |tag: &str| args.tags.iter().any(|s| tag == s || tag.starts_with(&format!("{s}/")));
-
-    let mut results: Vec<(Note, Vec<String>, Vec<LocatedTag>)> = notes
-        .into_iter()
-        .filter_map(|note| {
-            let fm_matches: Vec<String> = note.tags.iter().filter(|t| tag_matches_search(t)).cloned().collect();
-            let inline_matches: Vec<LocatedTag> = note
-                .inline_tags()
-                .into_iter()
-                .filter(|lt| tag_matches_search(&lt.tag))
-                .collect();
-            if fm_matches.is_empty() && inline_matches.is_empty() {
-                None
-            } else {
-                Some((note, fm_matches, inline_matches))
-            }
-        })
-        .collect();
-    sort_notes_by(&mut results, |(n, _, _)| &n.path, &args.sort);
-
-    match args.format {
-        OutputFormat::Plain => output::print_tags_search_plain(&results, &vault.path),
-        OutputFormat::Json => output::print_tags_search_json(&results, &vault.path),
-    }
-    Ok(())
-}
-
-fn cmd_tags_list(vault: Vault, args: TagsListArgs) -> eyre::Result<()> {
-    let notes: Vec<Note> = vault.search().execute()?.into_iter().filter_map(|r| r.ok()).collect();
-
-    let mut all_tags: std::collections::BTreeSet<String> = std::collections::BTreeSet::new();
-    for note in &notes {
-        for tag in &note.tags {
-            all_tags.insert(tag.clone());
-        }
-        for lt in note.inline_tags() {
-            all_tags.insert(lt.tag);
-        }
-    }
-    let tags: Vec<String> = all_tags.into_iter().collect();
-
-    match args.format {
-        OutputFormat::Plain => output::print_tags_list_plain(&tags),
-        OutputFormat::Json => output::print_tags_list_json(&tags),
-    }
-    Ok(())
-}
-
-fn update_single_note(vault: &Vault, note_arg: &std::path::Path, tags: &[String]) -> eyre::Result<Note> {
-    let (note_path, _) = resolve_note_path(vault, &note_arg.to_path_buf())?;
-    let mut note = Note::from_path(&note_path)?;
-
-    let new_tags: Vec<String> = tags.iter().filter(|t| !note.tags.contains(*t)).cloned().collect();
-    if !new_tags.is_empty() {
-        let fm = note.frontmatter.get_or_insert_with(indexmap::IndexMap::new);
-        let tags_entry = fm
-            .entry("tags".to_string())
-            .or_insert_with(|| gray_matter::Pod::Array(Vec::new()));
-        if let gray_matter::Pod::Array(arr) = tags_entry {
-            for tag in &new_tags {
-                arr.push(gray_matter::Pod::String(tag.clone()));
-            }
-        }
-        note.tags.extend(new_tags);
-        note.write()?;
-    }
-
-    Ok(note)
-}
-
-fn cmd_note_update(vault: Vault, args: UpdateArgs) -> eyre::Result<()> {
-    if let Some(note_path) = args.note {
-        let note = update_single_note(&vault, &note_path, &args.tag)?;
-        match args.format {
-            OutputFormat::Plain => output::print_note_update_plain(&note, &vault.path),
-            OutputFormat::Json => output::print_note_update_json(&note, &vault.path),
-        }
-        return Ok(());
-    }
-
-    if std::io::stdin().is_terminal() {
-        eyre::bail!("no note path provided and stdin is a TTY");
-    }
-
-    let stdin = std::io::stdin();
-    let mut notes = Vec::new();
-    for line in stdin.lock().lines() {
-        let line = line?;
-        let line = line.trim().to_string();
-        if line.is_empty() {
-            continue;
-        }
-        let note = update_single_note(&vault, std::path::Path::new(&line), &args.tag)?;
-        notes.push(note);
-    }
-
-    match args.format {
-        OutputFormat::Plain => {
-            for note in &notes {
-                output::print_note_update_plain(note, &vault.path);
-            }
-        }
-        OutputFormat::Json => output::print_note_update_many_json(&notes, &vault.path),
-    }
-    Ok(())
-}
+use args::{Cli, Command};
 
 fn main() -> eyre::Result<()> {
     color_eyre::install()?;
@@ -254,15 +23,15 @@ fn main() -> eyre::Result<()> {
     }
     let vault = Vault::open(&cli.vault)?;
     match cli.command {
-        Command::Search(args) => cmd_search(vault, args),
+        Command::Search(args) => search::cmd_search(vault, args),
         Command::Note(note_args) => match note_args.subcommand {
-            args::NoteCommand::Backlinks(args) => cmd_backlinks(vault, args),
-            args::NoteCommand::Rename(args) => cmd_rename(vault, args),
-            args::NoteCommand::Update(args) => cmd_note_update(vault, args),
+            args::NoteCommand::Backlinks(args) => note::cmd_backlinks(vault, args),
+            args::NoteCommand::Rename(args) => note::cmd_rename(vault, args),
+            args::NoteCommand::Update(args) => note::cmd_note_update(vault, args),
         },
         Command::Tags(tags_args) => match tags_args.subcommand {
-            args::TagsCommand::Search(args) => cmd_tags_search(vault, args),
-            args::TagsCommand::List(args) => cmd_tags_list(vault, args),
+            args::TagsCommand::Search(args) => tags::cmd_tags_search(vault, args),
+            args::TagsCommand::List(args) => tags::cmd_tags_list(vault, args),
         },
     }
 }
